Add python type annotations for the `execution_contexts` package.
This allows removing the runtime type checks and promote those failures from runtime to analysis time.

Note: not all types are annotated. Un-annotated types are assumed to be `Any`. This change only annotates enough types to safely remove the runtime checks in order to reduce the scope of this change.
PiperOrigin-RevId: 739265489